refactor(tui::ui::components::popups::message): move function "*mount_message" to popup file
