+++ tui/src/ui/components/mod.rs
@@ -52,9 +52,7 @@ pub use lyric::Lyric;
 pub use music_library::MusicLibrary;
 pub use playlist::Playlist;
 pub use podcast::{EpisodeList, FeedsList};
-pub use popups::{
-    HelpPopup, MessagePopup, PodcastAddPopup, SavePlaylistConfirmPopup, SavePlaylistPopup,
-};
+pub use popups::{HelpPopup, PodcastAddPopup, SavePlaylistConfirmPopup, SavePlaylistPopup};
 pub use progress::Progress;
 pub use tag_editor::*;
 use termusicplayback::SharedSettings;
