+++ tui/src/ui/components/popups/mod.rs
@@ -13,6 +13,7 @@ pub use deleteconfirm::{DeleteConfirmInputPopup, DeleteConfirmRadioPopup};
 #[allow(unused_imports)]
 pub use error::ErrorPopup;
 pub use help::HelpPopup;
+#[allow(unused_imports)]
 pub use message::MessagePopup;
 #[allow(unused_imports)]
 pub use podcast::{
